texk/dvipdfm-x: Minor bug fix and cleanup
git-svn-id: svn://tug.org/texlive/trunk/Build/source/texk/dvipdfm-x@35246 c570f23f-e606-0410-a88d-b1316a301751

Conflicts:
	bmpimage.c
	bmpimage.h
	jp2image.c
	jp2image.h

+++ xbb.c
@@ -166,7 +166,7 @@ static void write_xbb(char *fname,
 
 static void do_bmp (FILE *fp, char *filename)
 {
-  long   width, height;
+  int    width, height;
   double xdensity, ydensity;
 
   if (texpdf_bmp_get_bbox(fp, &width, &height, &xdensity, &ydensity) < 0) {
@@ -194,7 +194,7 @@ static void do_jpeg (FILE *fp, char *filename)
 
 static void do_jp2 (FILE *fp, char *filename)
 {
-  long   width, height;
+  int    width, height;
   double xdensity, ydensity;
 
   if (texpdf_jp2_get_bbox(fp, &width, &height, &xdensity, &ydensity) < 0) {
